Create grid layout for entire webpage
The layout for the entire page can be represented by a grid layout
of two columns, one column for the navigation and one column for
the dashboard itself. This structure has been added, with the dash-
board being four times as wide as the navigation column. The two
columns have been given a background color to visualize them.