Create structure for first row of header
The header section needs to be built with grid structures. The two
rows of the header do not line up, so the whole header cannot be
represented by one grid. Instead, each row is its own grid. The
first row grid has two columns and the left column is itself a
flex container. Being a flex container makes it easier for the search
bar to grow and shrink as needed to fill up space. However, the
second column is itself a grid container.