Add box shadow to header
The header needed a box shadow on the bottom edge to make it stand
out from the dashboard. The box shadow has been added and the z-index
of each element of the page has been changed so that the box shadow
only appears at the bottom edge of the header and not the left edge.